Refactor: flyte transactional funksjon slettVirksomhetTilstand
