+++ src/main/kotlin/no/nav/lydia/tilstandsmaskin/sideeffect/AngreVurderVirksomhetSideEffect.kt
@@ -18,7 +18,7 @@ import no.nav.lydia.tilstandsmaskin.lagreHendelse
 import no.nav.lydia.tilstandsmaskin.nyHendelseBasertPåSak
 import no.nav.lydia.tilstandsmaskin.settSakTilSlettet
 import no.nav.lydia.tilstandsmaskin.sideeffect.transactional.TilstandVirksomhetTransactional.Companion.oppdaterVirksomhetTilstand
-import no.nav.lydia.tilstandsmaskin.slettVirksomhetTilstand
+import no.nav.lydia.tilstandsmaskin.sideeffect.transactional.TilstandVirksomhetTransactional.Companion.slettVirksomhetTilstand
 import no.nav.lydia.tilstandsmaskin.tilVirksomhetIATilstand
 import no.nav.lydia.tilstandsmaskin.tilstand.VirksomhetKlarTilVurdering
 import org.slf4j.Logger
